Refactor ws_client and fix the timeout bug.
The nodemcu api, timeout is no longer an argument when making a tcp connection. This needs
to be improved in the future, for now it's just passing the correct argument.
Use compiled versions of the auxiliar files to improve memory consumption - I think having
everything in one file might improve it more...